Use more efficient deisgn pattern.
