+++ packages/core/src/index.ts
@@ -21,7 +21,6 @@ export * from './core/nonInteractiveToolExecutor.js';
 
 export * from './code_assist/codeAssist.js';
 export * from './code_assist/oauth2.js';
-export * from './code_assist/errors.js';
 
 // Export utilities
 export * from './utils/paths.js';
